requirements.txt
matplotlib==3.7.0
numpy==1.24.2
pandas==1.5.3
openpyxl==3.1.2
wordcloud==1.9.2
pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0
